refactor: Changed to matrices of variables x,u
Modified to be more similiar to the distributed version. Instead of a
single vector with all xi(t) and ui(t), now we have two matrices for each
agent, one for x and another for u, of size n-by-H and p-by-H, resp.
Changed the centralised code to work with this modifications.